+++ CHANGELOG.md
@@ -8,8 +8,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 
 ### TODO
-
 - Add tests
+- Add scaffold config for failing managers
 
 ### Added
 
@@ -18,6 +18,28 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 
 ### Removed
+
+
+## [0.9.4] - 2024-02-22
+
+### Added
+- Git: Possility to add :latest flag to always pull the latest commit from main
+- DNF: Versions in list function
+- Enhanced terminal printer
+- Github Manager
+
+### Fixed
+- Go Manager: Bug when bin folder did not exist 
+
+### Changed
+- Go updated to v1.22.0
+
+
+## [0.9.3] - 2024-02-01
+
+### Fixed
+
+- Bugfix in DNF COPR 
 
 
 ## [0.9.2] - 2023-11-23
@@ -44,7 +66,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Initial release
 
-[unreleased]: https://github.com/lucas-ingemar/packtrak/compare/v0.9.2...HEAD
+[unreleased]: https://github.com/lucas-ingemar/packtrak/compare/v0.9.4...HEAD
+[0.9.4]: https://github.com/lucas-ingemar/packtrak/compare/v0.9.3...v0.9.4
+[0.9.3]: https://github.com/lucas-ingemar/packtrak/compare/v0.9.2...v0.9.3
 [0.9.2]: https://github.com/lucas-ingemar/packtrak/compare/v0.9.1...v0.9.2
 [0.9.1]: https://github.com/lucas-ingemar/packtrak/compare/v0.9.0...v0.9.1
 [0.9.0]: https://github.com/lucas-ingemar/packtrak/releases/tag/v0.9.0